+++ src/main/java/com/psj/itembrowser/order/domain/vo/OrdersProductRelationResponseDTO.java
@@ -20,30 +20,30 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 public class OrdersProductRelationResponseDTO {
-
+	
 	@NotNull
 	Long groupId;
-
+	
 	@NotNull
 	Long productId;
-
+	
 	ProductResponseDTO productResponseDTO;
-
+	
 	@NotNull
 	@PositiveOrZero
-	Integer productQuantity;
-
+	Long productQuantity;
+	
 	LocalDateTime createdDate;
 	LocalDateTime updatedDate;
 	LocalDateTime deletedDate;
-
+	
 	public static OrdersProductRelationResponseDTO from(OrdersProductRelation ordersProductRelation) {
 		if (ordersProductRelation == null) {
 			return null;
 		}
-
+		
 		ProductResponseDTO productResponseDTO = ProductResponseDTO.from(ordersProductRelation.getProduct());
-
+		
 		return new OrdersProductRelationResponseDTO(
 			ordersProductRelation.getGroupId(),
 			ordersProductRelation.getProductId(),
@@ -54,14 +54,14 @@ public class OrdersProductRelationResponseDTO {
 			ordersProductRelation.getDeletedDate()
 		);
 	}
-
+	
 	public static OrdersProductRelationResponseDTO from(OrdersProductRelationEntity entity) {
 		if (entity == null) {
 			return null;
 		}
-
+		
 		ProductResponseDTO productResponseDTO = ProductResponseDTO.from(entity.getProduct());
-
+		
 		return new OrdersProductRelationResponseDTO(
 			entity.getGroupId(),
 			entity.getProductId(),
